src/modules/api/pipeline-task/requirements.txt
Flask==2.0.2
gcloud==0.18.3
google_cloud_secret_manager==2.7.2
google-api-python-client==2.24.0
google-cloud-datastore==2.2.0
gunicorn==20.1.0
python-dotenv==0.19.1